requirements.txt
arrow==1.2.3
asgiref==3.7.2
blessed==1.20.0
Django==4.2.2
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
inflection==0.5.1
packaging==23.1
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==3.5.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
wcwidth==0.2.6
